+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require_relative './base_decorator'
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
