--- trimmerDecorator.rb
@@ -1,13 +1,0 @@
-require_relative './baseDecorator'
-
-class TrimmerDecorator < BaseDecorator
-    def correct_name
-        length = @nameable.correct_name.length
-        if length <= 10
-            @nameable.correct_name
-        else
-            @nameable.correct_name[0..10]
-        end
-    end
-end
-
